requirements.txt
django==3.0.5
psycopg2==2.8.6
djangorestframework
celery==4.4.1
redis==3.2.0
django-filter